% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airportscreening-package.R
\docType{package}
\name{airportscreening-package}
\alias{airportscreening}
\alias{airportscreening-package}
\title{airportscreening: Run the CMMID Airport Screening Outcomes Shiny App}
\description{
Runs a Shiny app developed by the Centre for the Mathematical Modelling of Infectious Diseases (CMMID) to model the outcomes of implementing airport screening for potentially infected air passengers, given certain pathogen and air travel parameters, such as the time between infection and symptom onset, and the flight duration. This app was developed as part of the CMMID response to the Covid-19 pandemic, in early 2020.
}
\author{
\strong{Maintainer}: Billy Quilty \email{billy.quilty@lshtm.ac.uk} [copyright holder]

Authors:
\itemize{
  \item Sam Clifford \email{Sam.Clifford@lshtm.ac.uk} [copyright holder]
  \item Stefan Flasche \email{Stefan.Flasche@lshtm.ac.uk} [copyright holder]
  \item Rosalid Eggo \email{Rosalind.Eggo@lshtm.ac.uk} [copyright holder]
}

}
\keyword{internal}
