% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_histories}
\alias{generate_histories}
\title{Simulate travel and infection histories from user input data}
\usage{
generate_histories(
  dur.flight,
  mu_inc,
  sigma_inc,
  mu_inf,
  sigma_inf,
  sens.exit,
  sens.entry,
  prop.asy,
  sims
)
}
\arguments{
\item{dur.flight}{The flight duration in hours.}

\item{mu_inc}{Mean incubation period in days.}

\item{sigma_inc}{Variance of incubation period in days.}

\item{mu_inf}{Mean time to symptom onset in days.}

\item{sigma_inf}{Variance in time to symptom onset, in days.}

\item{sens.exit}{Sensitivity of tests used upon departure.}

\item{sens.entry}{Sensitivity of tests used upon arrival.}

\item{prop.asy}{Proportion of asymptomatic infections.}

\item{sims}{Number of simulation runs.}
}
\value{
A data.frame of travel and infection outcomes.
}
\description{
Simulate travel and infection histories from user input data
}
\keyword{internal}
