% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_probabilities}
\alias{generate_probabilities}
\title{Work out the detection probabilities of travellers}
\usage{
generate_probabilities(travellers)
}
\arguments{
\item{travellers}{Output from the \code{\link[=generate_travellers]{generate_travellers()}} function.}
}
\value{
A data.frame giving the probabilities of travellers who are infected
being detected as such at different stages of airline travel.
}
\description{
Work out the detection probabilities of travellers
}
\keyword{internal}
