% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_travellers}
\alias{generate_travellers}
\title{Generate travellers to have screening applied}
\usage{
generate_travellers(input, i)
}
\arguments{
\item{input}{Input from the Shiny app, giving the duration of the flight,
the mean incubation period, the variance of the incubation period, the mean
time between infection and symptom onset, the variance of the time to symptom
onset, the sensitivity of testing upon departure, the sensitivity of testing
upon arrival, and the proportion of asymptomatic infections.}

\item{i}{The number of simulation runs.}
}
\value{
A data.frame giving the proportion of travellers who are symptomatic
upon arrival and departure, given the pathogen parameters and flight duration
and the proportions that have severe infections upon arrival, and also the
proportion which is infected but undetected upon arrival.
}
\description{
Generate travellers to have screening applied
}
\keyword{internal}
