% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_stamp.R
\name{get_date_stamp}
\alias{get_date_stamp}
\title{Prepare a date and attribution statement}
\usage{
get_date_stamp()
}
\value{
A two element vector giving the date and attribution, and a link to
the referenced preprint.
}
\description{
Prepare a date and attribution statement
}
\keyword{internal}
