% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_ci_label}
\alias{make_ci_label}
\title{Make confidence interval labels}
\usage{
make_ci_label(x)
}
\arguments{
\item{x}{A vector of the central estimate, and the lower and upper confidence
limits.}
}
\value{
A string vector of the format `"I (I, I)"``.
}
\description{
Make confidence interval labels
}
\keyword{internal}
