% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{moment_match}
\alias{moment_match}
\title{Convert mean and variance to shape and rate for a gamma distribution}
\usage{
moment_match(mean, var)
}
\arguments{
\item{mean}{The mean of the gamma distribution from which to draw waiting
times.}

\item{var}{The variance of gamma distribution from which to draw waiting
times.}
}
\value{
A named list, where \code{alpha} is the \code{shape} of the gamma distribution,
while \code{beta} is the \code{rate}.
}
\description{
Convert mean and variance to shape and rate for a gamma distribution
}
\keyword{internal}
