% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_parameters.R
\docType{data}
\name{pathogen_parameters}
\alias{pathogen_parameters}
\title{Pathogen parameters taken from the literature.}
\format{
\subsection{\code{pathogen_parameters}}{

A data frame with 5 rows and 6 columns:
\describe{
\item{name}{The pathogen's (working) name}
\item{mu_inc}{The mean incubation period in days}
\item{sigma_inc}{The standard deviation of the incubation period, in days}
\item{mu_inf}{The mean time to symptom onset period in days}
\item{sigma_inf}{The standard deviation of the time to symptom onset, days}
\item{prop.asy}{The proportion of aymptomatic cases}
}
}
}
\source{
Multiple sources; see \code{data-raw/pathogen_parameters.R} for details.
}
\usage{
pathogen_parameters
}
\description{
Pathogen parameters taken from the literature.
}
\keyword{datasets}
