% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{CMMID Shiny for the efficacy of airport screening for infectious diseases}
\usage{
run_app()
}
\value{
Runs the \code{{shiny}} app locally.
}
\description{
Runs a local instance of a Shiny app developed by the Centre for
the Mathematical Modelling of Infectious Diseases at the London School of
Hygiene and tropical Medicine, to model the efficacy of implementing airport
screening for infectious diseases, as a part of the CMMID response to the
Covid-19 pandemic, in early 2020.
}
\examples{
# Choose options from the app GUI
\dontrun{
run_app()
}
}
