% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_to_event}
\alias{time_to_event}
\title{Draw waiting times until an event occurs}
\usage{
time_to_event(n, mean, var)
}
\arguments{
\item{n}{The number of waiting times to draw.}

\item{mean}{The mean of the gamma distribution from which to draw waiting
times.}

\item{var}{The variance of gamma distribution from which to draw waiting
times.}
}
\value{
A vector of length \code{n}, of the waiting times, or of the provided
mean repeated \eqn{n} times.
}
\description{
Draws a waiting time from either a gamma distribution
(\code{\link[=rgamma]{rgamma()}}), or provides a repeating sequence of the mean provided, if the
variance is 0.0.
}
\keyword{internal}
